+++ micro-ui/web/micro-ui-internals/packages/modules/core/CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog 
 
+## [1.8.32]  [10-Mar-2025]
+- FEATURE/DUCE-0246 :Userprofile password trimming
+
+## [1.8.31]  [10-Mar-2025]
+- FEATURE/DUCE-246 :Login page username and password trimming 
+
+## [1.8.30]  [07-Mar-2025]
+- BUGFIX/DUCE-243 : Updated breadcrumb and regex validation mapping in user profile
+
 ## [1.8.28]  [27-Feb-2025]
 - BUGFIX/SN-162 : Citizen sms otp fix